[BE] rename: 레포지토리 테스트 패키지 생성
