+++ BE/src/test/java/com/newsainturtle/shadowmate/yn/planner_setting/repository/CategoryColorRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.yn.planner_setting;
+package com.newsainturtle.shadowmate.yn.planner_setting.repository;
 
 import com.newsainturtle.shadowmate.planner_setting.entity.CategoryColor;
 import com.newsainturtle.shadowmate.planner_setting.repository.CategoryColorRepository;
